+++ docs/how_to_run.md
@@ -1,0 +1,20 @@
+# How to run
+
+
+## With config file
+
+go run main.go --config-file config.yaml
+
+### Config for StrategryBuyEveryDayIfBelowAverage strategry
+
+```yaml
+BuyEveryDayIfBelowAverage:
+  # Nongsim holdings
+  - code: 072710
+    quantity: 1
+  # Woori financial
+  - code: 316140
+    quantity: 4
+```
+
+## Without config file
