+++ internal/options/options.go
@@ -20,6 +20,9 @@ type Options struct {
 	KoreaInvestmentSecret  string `yaml:"koreainvestment_appsecret"`
 	KoreaInvestmentAccount string `yaml:"koreainvestment_account"`
 
+	// filled out when the config file has relevant configs.
+	BuyEveryDayIfBelowAverageConfig BuyEveryDayIfBelowAverageConfig
+
 	ConfigFile string
 
 	cmd *cobra.Command
